requirements.txt
beautifulsoup4==4.11.1
kafka==1.3.5
kafka_python==2.0.2
pandas==1.5.2
python-dotenv==0.21.0
requests==2.28.1
selenium==4.7.2
